Nihar has commited

+++ src/structs/rings/mod.rs
@@ -1,7 +1,7 @@
 // This rust module contains all our rings etc.
 // Some of these O(1) operations about multiplication and addition can be optimized a little bit
 // The DOmega ring can be stored directly without any underlying Dyad implentation, I feel
-pub type Int = i32;
+pub type Int = i64;
 pub type Float = f64;
 
 // Supertrait
